More succinct types

+++ clink/terminal/src/win_screen_buffer.h
@@ -12,20 +12,20 @@ class WinScreenBuffer
 public:
     virtual void    begin() override;
     virtual void    end() override;
-    virtual void    write(const char* data, int length) override;
+    virtual void    write(const char* data, int32 length) override;
     virtual void    flush() override;
-    virtual int     get_columns() const override;
-    virtual int     get_rows() const override;
+    virtual int32   get_columns() const override;
+    virtual int32   get_rows() const override;
     virtual void    clear(ClearType Type) override;
     virtual void    clear_line(ClearType Type) override;
-    virtual void    set_cursor(int column, int row) override;
-    virtual void    move_cursor(int dx, int dy) override;
-    virtual void    insert_chars(int count) override;
-    virtual void    delete_chars(int count) override;
+    virtual void    set_cursor(int32 column, int32 row) override;
+    virtual void    move_cursor(int32 dx, int32 dy) override;
+    virtual void    insert_chars(int32 count) override;
+    virtual void    delete_chars(int32 count) override;
     virtual void    set_attributes(const Attributes attr) override;
 
 private:
-    enum : unsigned short
+    enum : uint16
     {
         attr_mask_fg        = 0x000f,
         attr_mask_bg        = 0x00f0,
@@ -35,7 +35,7 @@ private:
     };
 
     void*           _handle = nullptr;
-    unsigned long   _prev_mode = 0;
-    unsigned short  _default_attr = 0x07;
+    uint32          _prev_mode = 0;
+    uint16          _default_attr = 0x07;
     bool            _bold = false;
 };
